+++ app/src/main/kotlin/ImageTheming/Main.kt
@@ -1,3 +1,7 @@
+package ImageTheming
+
+import ImageTheming.ImgController;
+
 fun main(args: Array<String>) {
     try {
         ImgController(args)
